+++ idea-plugin/src/main/kotlin/arrow/meta/ide/plugins/external/ui/tooltip/formatting/Styles.kt
@@ -1,13 +1,13 @@
-package arrow.meta.ide.plugins.initial.tooltip.formatting
+package arrow.meta.ide.plugins.external.ui.tooltip.formatting
 
 import com.intellij.util.ui.Html
 import com.intellij.util.ui.StartupUiUtil
 import com.intellij.util.ui.UIUtil
 
-const val cdnStyleLight = "https://47deg-academy.s3.amazonaws.com/css/edutools-style-light.css"
-const val cdnStyleDark = "https://47deg-academy.s3.amazonaws.com/css/edutools-style-dark.css"
+internal const val cdnStyleLight = "https://47deg-academy.s3.amazonaws.com/css/edutools-style-light.css"
+internal const val cdnStyleDark = "https://47deg-academy.s3.amazonaws.com/css/edutools-style-dark.css"
 
-fun Html.applyStylesFromCDN(): String {
+internal fun Html.applyStylesFromCDN(): String {
   val htmlBody = UIUtil.getHtmlBody(this)
   val stylesToUse = if (StartupUiUtil.isUnderDarcula()) {
     cdnStyleLight
